# LPS-148445 Webpack based build MVP