--- a/modules/apps/frontend-js/frontend-js-loader-modules-extender/package.json
+++ b/modules/apps/frontend-js/frontend-js-loader-modules-extender/package.json
@@ -1,6 +1,6 @@
 {
 	"dependencies": {
-		"@liferay/amd-loader": "4.3.1"
+		"@liferay/amd-loader": "5.0.0"
 	},
 	"name": "frontend-js-loader-modules-extender",
 	"version": "6.0.18"