--- a/modules/apps/frontend-js/frontend-js-react-web/npmscripts.config.js
+++ b/modules/apps/frontend-js/frontend-js-react-web/npmscripts.config.js
@@ -19,7 +19,7 @@ module.exports = {
 			'formik',
 			{name: 'prop-types', symbols: 'auto'},
 			{name: 'react', symbols: 'auto'},
-			'react-dnd',
+			{name: 'react-dnd', symbols: 'auto' },
 			'react-dnd-html5-backend',
 			{name: 'react-dom', symbols: 'auto'},
 		],