--- a/modules/apps/frontend-js/frontend-js-react-web/src/main/java/com/liferay/frontend/js/react/web/internal/importmap/FrontendJSReactWebJSImportmapsContributor.java
+++ b/modules/apps/frontend-js/frontend-js-react-web/src/main/java/com/liferay/frontend/js/react/web/internal/importmap/FrontendJSReactWebJSImportmapsContributor.java
@@ -47,7 +47,7 @@ public class FrontendJSReactWebJSImportmapsContributor
 			_importmapsJSONObject.put(
 				moduleName,
 				StringBundler.concat(
-					contextPath, "/__liferay__/amd2esm/", moduleName, ".js"));
+					contextPath, "/__liferay__/exports/", moduleName, ".js"));
 		}
 	}