--- a/modules/apps/frontend-js/frontend-js-react-web/src/main/resources/META-INF/resources/js/index.ts
+++ b/modules/apps/frontend-js/frontend-js-react-web/src/main/resources/META-INF/resources/js/index.ts
@@ -14,6 +14,8 @@
 
 import process from 'process';
 
+export {default as ReactDOMServer} from 'react-dom/server';
+
 export {default as render} from './render';
 export {default as ReactPortal} from './ReactPortal';
 export {default as useEventListener} from './hooks/useEventListener';