--- a/modules/apps/frontend-js/frontend-js-react-web/types/src/main/resources/META-INF/resources/js/index.d.ts
+++ b/modules/apps/frontend-js/frontend-js-react-web/types/src/main/resources/META-INF/resources/js/index.d.ts
@@ -12,6 +12,7 @@
  * details.
  */
 
+export {default as ReactDOMServer} from 'react-dom/server';
 export {default as render} from './render';
 export {default as ReactPortal} from './ReactPortal';
 export {default as useEventListener} from './hooks/useEventListener';