--- a/modules/apps/frontend-js/frontend-js-state-web/npmscripts.config.js
+++ b/modules/apps/frontend-js/frontend-js-state-web/npmscripts.config.js
@@ -14,8 +14,6 @@
 
 module.exports = {
 	build: {
-		bundler: {
-			ignore: ['**/legacy/config.js'],
-		},
+		main: 'src/main/resources/META-INF/resources/index.ts',
 	},
 };