--- a/modules/apps/frontend-js/frontend-js-web/package.json
+++ b/modules/apps/frontend-js/frontend-js-web/package.json
@@ -1,6 +1,6 @@
 {
 	"dependencies": {
-		"@liferay/amd-loader": "4.3.1",
+		"@liferay/amd-loader": "5.0.0",
 		"clay": "2.18.1",
 		"lodash.escape": "4.0.1",
 		"lodash.groupby": "4.6.0",