--- a/modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/undo/UndoHistory.js
+++ b/modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/undo/UndoHistory.js
@@ -14,7 +14,6 @@
 
 import {ClayButtonWithIcon} from '@clayui/button';
 import ClayDropDown, {Align} from '@clayui/drop-down';
-import ClayDropDownDivider from '@clayui/drop-down/lib/Divider';
 import {
 	ReactPortal,
 	useEventListener,
@@ -97,7 +96,7 @@ export default function UndoHistory() {
 						type={UNDO_TYPES.undo}
 					/>
 
-					<ClayDropDownDivider />
+					<ClayDropDown.Divider />
 
 					<ClayDropDown.Item
 						disabled={!undoHistory.length}