--- a/modules/package.json
+++ b/modules/package.json
@@ -13,7 +13,10 @@
 	"private": true,
 	"resolutions": {
 		"@liferay/eslint-plugin": "1.2.0",
-		"ua-parser-js": "^0.7.30"
+		"@types/html-minifier-terser": "5.1.0",
+		"html-webpack-plugin": "5.5.0",
+		"ua-parser-js": "^0.7.30",
+		"webpack": "5.70.0"
 	},
 	"scripts": {
 		"checkFormat": "liferay-npm-scripts check",