--- a/modules/package.json
+++ b/modules/package.json
@@ -1,7 +1,7 @@
 {
 	"devDependencies": {
 		"@liferay/eslint-plugin": "1.2.0",
-		"@liferay/npm-scripts": "47.3.0-pre.3",
+		"@liferay/npm-scripts": "47.3.0-pre.4",
 		"acorn": "7.1.0",
 		"alloy-ui": "^3.1.0-deprecated.40",
 		"babel-plugin-incremental-dom": "3.4.0",