… modules/apps/frontend-taglib/frontend-taglib-clay/npmscripts.config.js
… modules/npmscripts.config.js
… modules/yarn.lock