Hey @brianchandotcom, this is the first PR to begin the migration from the old liferay-npm-bundler + AMD loader to a new architecture based on webpack + ES browser modules.

This only replaces the build in `frontend-js-react-web`, but I will be sending more PRs to migrate all modules in the portal.

The benefits of this shift in the build are:

1. We simplify the build a lot (no need to use liferay-npm-bundler and smaller npm-scripts)
2. We are walking toward more standard deployment (ESM are standard, as opposed to our AMD loader)
3. The build is way more faster
4. Better interoperation between JS artifacts due to the use of ESM. For example, any JS file you deploy to the portal (no matter if built with our tools or not) can access all our published APIs because ES syntax is standard. Before everything needed to pass through the AMD loader pipe.

Example of how build times improve. These are the figures for a `gradle clean deploy -a` inside `frontend-js-react-web`: 

```
Before:
  Total:    66s
  Bundling: 32s

After:
  Total:    39s
  Bundling: 2s
```

----

This originates in https://github.com/liferay-frontend/liferay-portal/pull/2080

And we are pretty sure that CI passes because:

1. We have run CI over this PR lots of times, while fixing and polishing it
2. The only failing tests are flakey

In any case, if something else breaks (because we don't have tests for it, for example), it should have a little impact, most probably. In other words: it wouldn't be a big breaking change, but a standard bug that we can address as any other bug.

Also, because this includes support to interoperate ESM with AMD, if something breaks heavily, we can easily rollback the PR or parts of it. I mean, we can move different parts of the portal from ESM to AMD and the other way round as we need.